Fix conflicts and use new testcase
